requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements > requirements.txt
#

-i https://pypi.org/simple
alembic==1.8.1
boto3==1.26.26
botocore==1.29.26
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1; python_version >= '3.6' and python_version < '4.0'
email-validator==1.3.0
flask-cors==3.0.10
flask-login==0.6.2
flask-migrate==4.0.0
flask-sqlalchemy==3.0.2
flask-wtf==1.0.1
flask==2.2.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
jinja2==3.1.2
jmespath==1.0.1
mako==1.2.4
markupsafe==2.1.1
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
requests===2.28.1
s3transfer==0.6.0
setuptools==65.6.3; python_version >= '3.7'
six==1.16.0
sqlalchemy==1.4.44
urllib3==1.26.13
werkzeug==2.2.2
wtforms[email]==3.0.1
zipp==3.11.0
